drugstore/requirements.txt
django == 3.1.2
djangorestframework == 3.12.1
django-filter
django-cors-headers
djangorestframework-simplejwt == 4.6.0
#django-silk == 0.5.2
